requirements.txt
fastapi
uvicorn
python-multipart
groq
# OCR & Image Processing
paddlepaddle==2.6.1
paddleocr>=2.7.0.3
opencv-python-headless
numpy
Pillow
# Ebook creation (keep for future)
EbookLib